file renaming
- Changed the name of paymentControll.ts to shopController.ts
