+++ backend/src/routes/paymentRoutes.ts
@@ -1,5 +1,5 @@
 import express from 'express';
-import { buyItem } from '../controllers/transaction_controllers/paymentController';
+import { buyItem } from '../controllers/transaction_controllers/shopController';
 import { handlePaymentCallback } from '../controllers/transaction_controllers/webhookController'; // Correct import path
 
 const router = express.Router();
